backend/requirements.txt
requests==2.31.0
pandas==2.1.4
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2
openpyxl==3.1.2
gigachat==0.1.9
python-dotenv==1.0.0

fastapi
uvicorn[standard]

aiogram

aiohttp
pydantic

fastapi==0.104.1

sqlite3
